Logging game state
At the start of a turn, and whenever a player ends turn, we call from
c++ into the python method getGameStateString in CvGameInterface.py and
get a parseable description of the current game state. This is logged to
disk by the c++ code. The game state can be used e.g. by an observer web
application.